Implement basic input and calculator functionality
Created a calculator object that takes a string and executes
mathmetical operators on the contents of the string.

Created event listners for all buttons and applicable key presses.
Event handler for both listener will either concat a new digit to
the input or execute a special operation such as creating a
calculator object and returning the mathematical result.